Add FakeBlobStorageService
This allows us to simulate the entire logic. The passing around of
folderPrefixes was making it difficult to ensure that the behaviour
demonstrated by the tests was accurate. By faking the blob storage
service using the existing IBlobStorageService we can get an accurate
test of the end-to-end behaviour without having to fake an actual azure
blob sdk.